refactor: OptionResponse dto 분리
