+++ src/main/java/com/ray/pominowner/menu/controller/dto/OptionGroupWithOptionResponse.java
@@ -1,10 +1,8 @@
 package com.ray.pominowner.menu.controller.dto;
 
-import com.ray.pominowner.menu.domain.Option;
 import com.ray.pominowner.menu.domain.OptionGroup;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public record OptionGroupWithOptionResponse(
         OptionGroupResponse optionGroupResponse,
@@ -22,28 +20,6 @@ public record OptionGroupWithOptionResponse(
     }
 
     private record OptionGroupResponse(String name, int maxOptionCount, boolean required) {
-    }
-
-    public record OptionResponse(String name, int price, boolean selected) {
-
-        public static List<OptionResponse> from(List<Option> options) {
-            return options.stream()
-                    .map(option -> new OptionResponse(
-                            option.getName(),
-                            option.getPrice(),
-                            option.isSelected()
-                    ))
-                    .collect(Collectors.toList());
-        }
-
-        public static OptionResponse from(Option option) {
-            return new OptionResponse(
-                    option.getName(),
-                    option.getPrice(),
-                    option.isSelected()
-            );
-        }
-
     }
 
 }
